Fix narrowing conversions and missing function declarations
1. Fix narrowing conversion warnings for chrono timestamps:
   - Added static_cast<uint64_t>() to convert long int to uint64_t
   - Line 346: record_gate_result() constructor initializer
   - Line 372: record_violation() struct initializer
   - Resolves [-Wnarrowing] warnings in chrono conversions

2. Uncomment function declarations that were previously commented:
   - llama_record_pcie_transfer(bool is_h2d, uint64_t size)
   - llama_record_allocation()
   - These functions are used in macros and need public declarations
   - Resolves [-Wmissing-declarations] warnings

+++ src/llama-decode-acceptance-criteria.cpp
@@ -343,7 +343,7 @@ bool decode_acceptance_validator::record_gate_result(acceptance_gate gate_id,
         description,
         passed,
         failure_reason,
-        std::chrono::high_resolution_clock::now().time_since_epoch().count(),
+        static_cast<uint64_t>(std::chrono::high_resolution_clock::now().time_since_epoch().count()),
         is_critical
     };
 
@@ -369,7 +369,7 @@ bool decode_acceptance_validator::record_violation(const char * description,
     acceptance_violation violation = {
         description,
         gate,
-        std::chrono::high_resolution_clock::now().time_since_epoch().count(),
+        static_cast<uint64_t>(std::chrono::high_resolution_clock::now().time_since_epoch().count()),
         evidence_source,
         is_blocking
     };
